NEW_PROV_BACKEND/fastapi-crud/src/requirements.txt
fastapi==0.109.2
psycopg2-binary==2.9.5
uvicorn==0.27.1
loguru
provisio
dongraphio

# dev
pytest==7.2.0
httpx==0.23.1